Update docs for toast api and force data refresh api
