+++ packages/ui-extensions/docs/surfaces/customer-account/reference/apis/toast.doc.ts
@@ -1,0 +1,21 @@
+import type {ReferenceEntityTemplateSchema} from '@shopify/generate-docs';
+
+import {CUSTOMER_ACCOUNT_STANDARD_API_DEFINITION} from '../helper.docs';
+
+const data: ReferenceEntityTemplateSchema = {
+  name: 'Toast',
+  description: 'The API for interacting with toast.',
+  isVisualComponent: false,
+  category: 'APIs',
+  type: 'API',
+  definitions: [
+    {
+      title: CUSTOMER_ACCOUNT_STANDARD_API_DEFINITION.title,
+      description: CUSTOMER_ACCOUNT_STANDARD_API_DEFINITION.description,
+      type: 'Docs_Standard_ToastApi',
+    },
+  ],
+  related: [],
+};
+
+export default data;
