+++ packages/ui-extensions/src/surfaces/customer-account/api/shared.ts
@@ -392,41 +392,6 @@ export interface CompanyLocation {
   id: string;
 }
 
-export interface Ui {
-  /**
-   * The Toast API displays a non-disruptive message that displays at the bottom
-   * of the interface to provide quick, at-a-glance feedback on the outcome
-   * of an action.
-   *
-   * How to use:
-   *
-   * - Use toasts to confirm successful actions.
-   *
-   * - Aim for two words.
-   *
-   * - Use noun + past tense verb format. For example, \`Changes saved\`.
-   *
-   * For errors, or information that needs to persist on the page, use a [banner](/docs/api/checkout-ui-extensions/unstable/components/feedback/banner) component.
-   */
-  toast: {
-    show(content: string): void;
-  };
-  /**
-   * Refresh data so the surrounding information on the page is updated. The `content` string will appear in a toast message after refresh, to confirm the action was successful.
-   *
-   * To request access to this API:
-   *
-   * 1. Go to your partner dashboard and click **Apps**.
-   *
-   * 2. Select the app you need to request access for.
-   *
-   * 3. Click **API access**.
-   *
-   * 4. Under **Access force data refresh**, click **Request access**.
-   */
-  forceDataRefresh(content: string): Promise<void>;
-}
-
 export interface SessionToken {
   /**
    * Requests a session token that hasn't expired. You should call this method every
